#66 - SingleCardMix implemented; still issues
The problemis that each home content section on touch converts from a multiline columns layout to a touch-friendly draggable row of content.

To visually inform the user that there's more content to see in these sections, the width of each content is two-fifths on touch, meaning that the third content item on page load is half cut-off, inviting the user to drag it fully into view.

On the mixes page, there is no draggable single row so on touch, each row has a one-fifth size space on the right since it's still a multiline arrangement.

Big changes needed to SingleMixCard to make it all work.