feat: add working OutsideClick base code to ScheduleBar
This is an inline component called inside ScheduleBar.js that, as of this test, displays an alert, 'You clicked outside the open schedule'. It doesn't show this alert for clicks inside the schedule bar or the open menu.

However, additional functionality is needed so that when you click on the Live button, search icon or chat icon, it closes the schedule bar.